Date object is changed.

+++ app/src/main/java/com/yilmazvolkan/simplebitcoinapp/data/database/DataEntity.kt
@@ -11,7 +11,7 @@ data class DataEntity(
     @ColumnInfo(name = "id")
     val id: Int = 0,
     @ColumnInfo(name = "timestamp")
-    val x: Int,
+    val x: Long,
     @ColumnInfo(name = "value")
     val y: Float
 )
