+++ app/src/main/java/com/yilmazvolkan/simplebitcoinapp/models/BitcoinData.kt
@@ -1,6 +1,6 @@
 package com.yilmazvolkan.simplebitcoinapp.models
 
 data class BitcoinData(
-    val x: Int,
+    val x: Long,
     val y: Float
 )
